Rectangle/ Lattice Multiplication
DESCRIPTION:
Source: Wikipedia
Rectangle/ Lattice multiplication as it is more commonly known is used to multiply two multi digit numbers in a faster way when compared to traditional way of multiplication. It is known by other names such as gelusia multiplication, sieve multiplication, shabakh, Venetian squares or Hindu lattice. Lattice multiplication tries to break the problem into several parts which makes it easier to find the product of two large numbers.

APPROACH:
In order to handle very large numbers, the concept of strings has been brought into the picture. Multiplication starts with the final character of 1st string to the 1st character. For the second string, it starts from 1st to the last character. The logic for executing this is done by traversing the lattice and storing the elements and adding them, so that all the nodes are visited and contributes to the final answer. Contrary to Ala Carte multiplication where long long int was used, this drawback has been overcome in Rectangle multiplication. String computations give the flexibility to choose very large numbers and compute the product in an efficient manner.
Further, this code is not best suited to compute product of real values.

RESULT:
The code is successfully tested for different categories of numbers which included positive, negative, zero and large numbers.